# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: Donald O. Oduol
    url: https://github.com/oduold
  description: Web inventory API for a supermarket store
  title: Brandlogs Supermarket API
  version: 0.0.1
servers:
- description: Development Server
  url: http://localhost:8080
...
paths:
...
  /transactions:
    get:
      description: List of StockTransactions Details
      operationId: stockTransactionsDetail
      parameters:
      - description: transaction type code
        explode: true
        in: query
        name: type
        required: false
        schema:
          type: string
        style: form
      - description: store code
        explode: true
        in: query
        name: store
        required: false
        schema:
          type: string
        style: form
      - description: Transaction Date from to date
        explode: true
        in: query
        name: since
        required: false
        schema:
          format: date
          type: string
        style: form
      - description: Transaction Date from
        explode: true
        in: query
        name: from
        required: false
        schema:
          format: date
          type: string
        style: form
      - description: Transaction Date to
        explode: true
        in: query
        name: to
        required: false
        schema:
          format: date
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TransactionDetail'
                type: array
          description: A paged array of StockTransaction Details
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Retrieve all StockTransactions Detail
      tags:
      - StockTransactionsDetails
      x-accepts: application/json
      x-tags:
      - tag: StockTransactionsDetails
    post:
      description: Add a StockTransaction
      operationId: addStockTransaction
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transaction'
      responses:
        "200":
          description: OK
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Add a StockTransaction
      tags:
      - StockTransactions
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: StockTransactions
  /transactions/{id}:
    get:
      description: StockTransaction data by id
      operationId: stockTransaction
      parameters:
      - description: resource id
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
          description: A StockTransaction's data
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Retrieve stock transaction data by id
      tags:
      - StockTransaction
      x-accepts: application/json
      x-tags:
      - tag: StockTransaction
  /transactiontypes:
    get:
      description: List of TransactionTypes
      operationId: transactionTypes
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TransactionType'
                type: array
          description: A paged array of TransactionTypes
          headers:
            x-next:
              description: A link to the next page of responses
              explode: false
              schema:
                type: string
              style: simple
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Retrieve all TransactionTypes
      tags:
      - TransactionTypes
      x-accepts: application/json
      x-tags:
      - tag: TransactionTypes
    post:
      description: Add a TransactionType
      operationId: addTransactionType
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionType'
      responses:
        "200":
          description: OK
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Add a TransactionType
      tags:
      - TransactionTypes
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: TransactionTypes
  /transactiontypes/{id}:
    get:
      description: TransactionType data by id
      operationId: transactionType
      parameters:
      - description: resource id
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionType'
          description: A TransactionType's data
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Retrieve transactionType data by id
      tags:
      - TransactionType
      x-accepts: application/json
      x-tags:
      - tag: TransactionType
    put:
      description: Update TransactionType data by id
      operationId: transactionTypeEdit
      parameters:
      - description: resource id
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionType'
        description: Update transactionType properties to be changed
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionType'
          description: A TransactionType's data
        default:
          $ref: '#/paths/~1items/get/responses/default'
      summary: Update transactionType data by id
      tags:
      - TransactionType
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: TransactionType
components:
  schemas:
    Item:
      example:
        vendor:
          code: code
          name: name
          id: 6
        name: name
        id: 0
        creationDate: 2000-01-23
      properties:
        id:
          format: int64
          type: integer
        name:
          type: string
        vendor:
          $ref: '#/components/schemas/Vendor'
        creationDate:
          format: date
          type: string
      type: object
    Location:
      example:
        code: code
        name: name
        id: 0
      properties:
        id:
          format: int64
          type: integer
        code:
          type: string
        name:
          type: string
      type: object
    Transaction:
      example:
        transactionType:
          code: code
          name: name
          id: 6
        transactionDetails:
        - null
        - null
        id: 0
        store:
          code: code
          name: name
          id: 0
        Vendor:
          code: code
          name: name
          id: 6
        transactionDate: 2000-01-23
      properties:
        id:
          format: int64
          type: integer
        transactionType:
          $ref: '#/components/schemas/TransactionType'
        store:
          $ref: '#/components/schemas/Location'
        Vendor:
          $ref: '#/components/schemas/Vendor'
        transactionDetails:
          items:
            $ref: '#/components/schemas/TransactionDetail'
          type: array
        transactionDate:
          format: date
          type: string
      type: object
    TransactionDetail:
      example:
        Item:
          vendor:
            code: code
            name: name
            id: 6
          name: name
          id: 0
          creationDate: 2000-01-23
        quantity: 1
        transaction:
          transactionType:
            code: code
            name: name
            id: 6
          transactionDetails:
          - null
          - null
          id: 0
          store:
            code: code
            name: name
            id: 0
          Vendor:
            code: code
            name: name
            id: 6
          transactionDate: 2000-01-23
      properties:
        transaction:
          $ref: '#/components/schemas/Transaction'
        Item:
          $ref: '#/components/schemas/Item'
        quantity:
          format: int32
          type: integer
      type: object
    TransactionType:
      example:
        code: code
        name: name
        id: 6
      properties:
        id:
          format: int64
          type: integer
        code:
          type: string
        name:
          type: string
      type: object
    Vendor:
      example:
        code: code
        name: name
        id: 6
      properties:
        id:
          format: int64
          type: integer
        code:
          type: string
        name:
          type: string
      type: object
...